Re-arranged Filesystem
All documentation has been placed in a documentation folder.
Some of the HTML file names were changed (as well as the links to them
in main.py).
CSS was removed from the HTML files.
Font-Awesome's files were placed in the corresponding /static/ folders.

I tried to remove all of the JS from the HTML, but ran into some
problems.  Maybe in the future.