Password encryption and login session handling
Updated so that passwords are encrypted. Logins are stored in SESSION
and checked in get_navbar and get_play_button_display. Currently, log in
is done at top of index - still need separate page for login. Logout
works, but it redirects to index, which immediately logs back in unless
you change credentials.